% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utils.R
\name{adapt_sql_for_platform}
\alias{adapt_sql_for_platform}
\title{Adapt SQL for Database Platform}
\usage{
adapt_sql_for_platform(sql, platform)
}
\arguments{
\item{sql}{SQL query string}

\item{platform}{Database platform}
}
\value{
Adapted SQL query
}
\description{
Adapt SQL for Database Platform
}
\keyword{internal}
