% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_analysis_db.R
\name{build_prevalence_sql}
\alias{build_prevalence_sql}
\title{Build SQL Query for Prevalence Analysis}
\usage{
build_prevalence_sql(
  cohort_database_schema,
  cdm_database_schema,
  cohort_table_name,
  concept_id_1,
  concept_id_2,
  domain,
  days_start,
  days_end
)
}
\arguments{
\item{cohort_database_schema}{Schema for cohort table}

\item{cdm_database_schema}{Schema for CDM tables}

\item{cohort_table_name}{Name of cohort table}

\item{concept_id_1}{Cohort concept ID}

\item{concept_id_2}{Test/procedure concept ID}

\item{domain}{OMOP domain (Procedure, Measurement, etc.)}

\item{days_start}{Start of time window}

\item{days_end}{End of time window}
}
\value{
Character SQL query
}
\description{
Build SQL Query for Prevalence Analysis
}
\keyword{internal}
