% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{cleanup_temp_tables}
\alias{cleanup_temp_tables}
\title{Clean Up Temporary Tables}
\usage{
cleanup_temp_tables(connection, schema, table_pattern = "temp_prevalence_\%")
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Character. Schema name}

\item{table_pattern}{Character. Pattern to match table names}
}
\description{
Clean Up Temporary Tables
}
