% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{create_connection_string}
\alias{create_connection_string}
\title{Create Database Connection String Helper}
\usage{
create_connection_string(
  server,
  database,
  trusted_connection = TRUE,
  uid = NULL,
  pwd = NULL
)
}
\arguments{
\item{server}{Character. Server name}

\item{database}{Character. Database name}

\item{trusted_connection}{Logical. Use Windows authentication}

\item{uid}{Character. Username (if not using trusted connection)}

\item{pwd}{Character. Password (if not using trusted connection)}
}
\value{
Character. Connection string
}
\description{
Create Database Connection String Helper
}
