% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_example_usage.R
\name{create_performance_indexes}
\alias{create_performance_indexes}
\title{Create Performance Indexes on Results Table}
\usage{
create_performance_indexes(connection, schema, table_name)
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Schema name}

\item{table_name}{Table name}
}
\description{
Create Performance Indexes on Results Table
}
