% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{create_results_index}
\alias{create_results_index}
\title{Create Index on Results Table}
\usage{
create_results_index(
  connection,
  schema,
  table_name,
  index_columns,
  index_name = NULL
)
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Character. Schema name}

\item{table_name}{Character. Table name}

\item{index_columns}{Character vector. Columns to index}

\item{index_name}{Character. Index name (optional)}
}
\description{
Create Index on Results Table
}
