% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utils.R
\name{create_results_indexes}
\alias{create_results_indexes}
\title{Create Indexes on Results Table}
\usage{
create_results_indexes(connection, schema, table_name)
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Schema name}

\item{table_name}{Table name}
}
\description{
Create Indexes on Results Table
}
\keyword{internal}
