% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_prevalence_analysis.R
\name{create_results_table_db}
\alias{create_results_table_db}
\title{Create Results Table in Database}
\usage{
create_results_table_db(connection, schema, table_name, overwrite = TRUE)
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Database schema}

\item{table_name}{Table name}

\item{overwrite}{Whether to overwrite existing table}
}
\description{
Create Results Table in Database
}
