% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_parser.R
\name{create_sample_tsv}
\alias{create_sample_tsv}
\title{Create Sample TSV Data}
\usage{
create_sample_tsv(output_path = "sample_analysis_data.tsv", n_rows = 5)
}
\arguments{
\item{output_path}{Character. Path where to save sample TSV file}

\item{n_rows}{Integer. Number of sample rows to generate}
}
\value{
Character. Path to created sample file
}
\description{
Create Sample TSV Data
}
