% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_example_usage.R
\name{create_sample_tsv_data}
\alias{create_sample_tsv_data}
\title{Create Sample TSV Data for Testing}
\usage{
create_sample_tsv_data(output_file, n_rows = 10)
}
\arguments{
\item{output_file}{Character. Output TSV file path}

\item{n_rows}{Integer. Number of sample rows to create}
}
\description{
Create Sample TSV Data for Testing
}
