% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_summary_stats}
\alias{create_summary_stats}
\title{Create Summary Statistics}
\usage{
create_summary_stats(results)
}
\arguments{
\item{results}{Data frame with prevalence analysis results}
}
\value{
Data frame with summary statistics
}
\description{
Create Summary Statistics
}
