% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_prevalence_analysis.R
\name{create_summary_table_db}
\alias{create_summary_table_db}
\title{Create Summary Statistics Table}
\usage{
create_summary_table_db(connection, schema, result_table_name)
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Schema name}

\item{result_table_name}{Results table name}
}
\description{
Create Summary Statistics Table
}
