% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_usage.R
\name{create_test_cohort_table}
\alias{create_test_cohort_table}
\title{Create Test Database with Sample Cohort Data}
\usage{
create_test_cohort_table(db_connection, cohort_table_name = "test_cohort")
}
\arguments{
\item{db_connection}{Database connection object}

\item{cohort_table_name}{Character. Name for the cohort table}
}
\description{
Create Test Database with Sample Cohort Data
}
