% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_usage.R
\name{demonstrate_prevalence_queries}
\alias{demonstrate_prevalence_queries}
\title{Demonstrate Query Functions}
\usage{
demonstrate_prevalence_queries(
  db_connection,
  results_table = "prevalence_results"
)
}
\arguments{
\item{db_connection}{Database connection object}

\item{results_table}{Character. Name of results table}
}
\description{
Demonstrate Query Functions
}
