% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utils.R
\name{drop_table}
\alias{drop_table}
\title{Drop Table if Exists}
\usage{
drop_table(connection, schema, table_name)
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Schema name}

\item{table_name}{Table name}
}
\description{
Drop Table if Exists
}
\keyword{internal}
