% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_prevalence_analysis.R
\name{execute_prevalence_query_db}
\alias{execute_prevalence_query_db}
\title{Execute Prevalence Query in Database}
\usage{
execute_prevalence_query_db(
  connection,
  cohort_schema,
  cdm_schema,
  cohort_table,
  domain_table,
  concept_id_2,
  days_start,
  days_end,
  cohortname
)
}
\arguments{
\item{connection}{Database connection}

\item{cohort_schema}{Cohort schema}

\item{cdm_schema}{CDM schema}

\item{cohort_table}{Cohort table name}

\item{domain_table}{Domain table name}

\item{concept_id_2}{Concept ID to search for}

\item{days_start}{Start of time window}

\item{days_end}{End of time window}

\item{cohortname}{Cohort name filter}
}
\value{
List with prevalence statistics
}
\description{
Execute Prevalence Query in Database
}
