% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_cohort_definitions}
\alias{get_cohort_definitions}
\title{Get Cohort Definition IDs and Names}
\usage{
get_cohort_definitions(connection, cohort_database_schema, cohort_table_name)
}
\arguments{
\item{connection}{Database connection}

\item{cohort_database_schema}{Character. Schema name}

\item{cohort_table_name}{Character. Cohort table name}
}
\value{
Data frame with cohort definitions
}
\description{
Get Cohort Definition IDs and Names
}
