% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_analysis.R
\name{get_cohort_patient_count}
\alias{get_cohort_patient_count}
\title{Get Patient Count from Cohort Table}
\usage{
get_cohort_patient_count(connection, cohort_table_name, cohortname)
}
\arguments{
\item{connection}{Database connection object}

\item{cohort_table_name}{Character. Name of the cohort table}

\item{cohortname}{Character. Name of the specific cohort}
}
\value{
Integer. Number of patients in the cohort
}
\description{
Get Patient Count from Cohort Table
}
