% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utils.R
\name{get_database_platform}
\alias{get_database_platform}
\title{Get Database Platform}
\usage{
get_database_platform(connection)
}
\arguments{
\item{connection}{Database connection}
}
\value{
Character string indicating database platform
}
\description{
Get Database Platform
}
\keyword{internal}
