% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_db_platform}
\alias{get_db_platform}
\title{Get Database Platform Information}
\usage{
get_db_platform(connection)
}
\arguments{
\item{connection}{Database connection}
}
\value{
Character. Database platform name
}
\description{
Get Database Platform Information
}
