% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_domain_concept_column}
\alias{get_domain_concept_column}
\title{Get Domain Concept Column Name}
\usage{
get_domain_concept_column(domain_table)
}
\arguments{
\item{domain_table}{Character. OMOP CDM table name}
}
\value{
Character. Concept ID column name for the table
}
\description{
Get Domain Concept Column Name
}
