% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_domain_date_column}
\alias{get_domain_date_column}
\title{Get Domain Date Column Name}
\usage{
get_domain_date_column(domain_table)
}
\arguments{
\item{domain_table}{Character. OMOP CDM table name}
}
\value{
Character. Date column name for the table
}
\description{
Get Domain Date Column Name
}
