% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_domain_table_name}
\alias{get_domain_table_name}
\title{Database Utility Functions for Prevalence Analysis}
\usage{
get_domain_table_name(domain)
}
\arguments{
\item{domain}{Character. OMOP domain name}
}
\value{
Character. Corresponding OMOP CDM table name
}
\description{
This file contains utility functions for database operations in prevalence analysis.
Get Domain Table Name from OMOP Domain
}
