% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_analysis.R
\name{get_patients_with_concept_in_timeframe}
\alias{get_patients_with_concept_in_timeframe}
\title{Get Patients with Specific Concept in Timeframe}
\usage{
get_patients_with_concept_in_timeframe(
  connection,
  cohort_table_name,
  cohortname,
  concept_id,
  domain,
  days_start,
  days_end,
  procedure_occurrence_table,
  measurement_table,
  condition_occurrence_table,
  drug_exposure_table,
  device_exposure_table,
  observation_table
)
}
\arguments{
\item{connection}{Database connection object}

\item{cohort_table_name}{Character. Name of the cohort table}

\item{cohortname}{Character. Name of the specific cohort}

\item{concept_id}{Integer. OMOP concept ID to search for}

\item{domain}{Character. OMOP domain (Procedure, Measurement, etc.)}

\item{days_start}{Integer. Start of time window relative to index date}

\item{days_end}{Integer. End of time window relative to index date}

\item{procedure_occurrence_table}{Character. Name of procedure occurrence table}

\item{measurement_table}{Character. Name of measurement table}

\item{condition_occurrence_table}{Character. Name of condition occurrence table}

\item{drug_exposure_table}{Character. Name of drug exposure table}

\item{device_exposure_table}{Character. Name of device exposure table}

\item{observation_table}{Character. Name of observation table}
}
\value{
Integer. Number of patients with the concept in the specified timeframe
}
\description{
Get Patients with Specific Concept in Timeframe
}
