% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_results.R
\name{get_results_by_cohort}
\alias{get_results_by_cohort}
\title{Get Results by Cohort}
\usage{
get_results_by_cohort(connection, cohort_database_schema, results_table_name)
}
\arguments{
\item{connection}{Database connection object}

\item{cohort_database_schema}{Character. Schema name where results table is stored}

\item{results_table_name}{Character. Name of the results table}
}
\value{
Data frame with results grouped by cohort
}
\description{
Get Results by Cohort
}
\examples{
\dontrun{
cohort_summary <- get_results_by_cohort(
  connection = connection,
  cohort_database_schema = "results",
  results_table_name = "prevalence_analysis_results"
)
}

}
