% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_results.R
\name{get_results_by_workflow_stage}
\alias{get_results_by_workflow_stage}
\title{Get Results by Workflow Stage}
\usage{
get_results_by_workflow_stage(
  connection,
  cohort_database_schema,
  results_table_name
)
}
\arguments{
\item{connection}{Database connection object}

\item{cohort_database_schema}{Character. Schema name where results table is stored}

\item{results_table_name}{Character. Name of the results table}
}
\value{
Data frame with results grouped by workflow stage
}
\description{
Get Results by Workflow Stage
}
\examples{
\dontrun{
workflow_summary <- get_results_by_workflow_stage(
  connection = connection,
  cohort_database_schema = "results",
  results_table_name = "prevalence_analysis_results"
)
}

}
