% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_results.R
\name{get_results_summary}
\alias{get_results_summary}
\title{Get Results Summary Statistics}
\usage{
get_results_summary(connection, cohort_database_schema, results_table_name)
}
\arguments{
\item{connection}{Database connection object}

\item{cohort_database_schema}{Character. Schema name where results table is stored}

\item{results_table_name}{Character. Name of the results table}
}
\value{
Data frame with summary statistics
}
\description{
Get Results Summary Statistics
}
\examples{
\dontrun{
summary_stats <- get_results_summary(
  connection = connection,
  cohort_database_schema = "results",
  results_table_name = "prevalence_analysis_results"
)
}

}
