% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_supported_domains}
\alias{get_supported_domains}
\title{Get Supported OMOP Domains}
\usage{
get_supported_domains()
}
\value{
Character vector of supported OMOP domains
}
\description{
Get Supported OMOP Domains
}
