% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utils.R
\name{get_table_row_count}
\alias{get_table_row_count}
\title{Get Table Row Count}
\usage{
get_table_row_count(connection, schema, table_name)
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Schema name}

\item{table_name}{Table name}
}
\value{
Integer row count
}
\description{
Get Table Row Count
}
\keyword{internal}
