% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utils.R, R/db_prevalence_analysis.R
\name{insert_results_to_db}
\alias{insert_results_to_db}
\title{Insert Results to Database}
\usage{
insert_results_to_db(
  connection,
  cohort_database_schema,
  result_table_name,
  results_data
)

insert_results_to_db(
  connection,
  cohort_database_schema,
  result_table_name,
  results_data
)
}
\arguments{
\item{connection}{Database connection}

\item{cohort_database_schema}{Schema name}

\item{result_table_name}{Table name}

\item{results_data}{Data frame with results}

\item{results_table_name}{Table name}
}
\description{
Insert Results to Database

Insert Results into Database Table
}
\keyword{internal}
