% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{optimize_sql_for_platform}
\alias{optimize_sql_for_platform}
\title{Optimize Query for Database Platform}
\usage{
optimize_sql_for_platform(sql, platform)
}
\arguments{
\item{sql}{Character. Base SQL query}

\item{platform}{Character. Database platform}
}
\value{
Character. Optimized SQL query
}
\description{
Optimize Query for Database Platform
}
