% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_parser.R
\name{parse_analysis_metadata}
\alias{parse_analysis_metadata}
\title{Parse Analysis Metadata from TSV}
\usage{
parse_analysis_metadata(tsv_data)
}
\arguments{
\item{tsv_data}{Data frame with TSV data}
}
\value{
Data frame with parsed metadata
}
\description{
Parse Analysis Metadata from TSV
}
\keyword{internal}
