% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_prevalence_analysis.R
\name{process_prevalence_chunk_db}
\alias{process_prevalence_chunk_db}
\title{Process Prevalence Analysis for Data Chunk}
\usage{
process_prevalence_chunk_db(
  chunk_data,
  connection,
  cohort_database_schema,
  cdm_database_schema,
  cohort_table_name
)
}
\arguments{
\item{chunk_data}{Data frame chunk to process}

\item{connection}{Database connection}

\item{cohort_database_schema}{Cohort schema}

\item{cdm_database_schema}{CDM schema}

\item{cohort_table_name}{Cohort table name}
}
\value{
Data frame with results
}
\description{
Process Prevalence Analysis for Data Chunk
}
