% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{table_exists_in_schema}
\alias{table_exists_in_schema}
\title{Check if Table Exists in Schema}
\usage{
table_exists_in_schema(connection, schema, table_name)
}
\arguments{
\item{connection}{Database connection}

\item{schema}{Character. Schema name}

\item{table_name}{Character. Table name}
}
\value{
Logical. TRUE if table exists
}
\description{
Check if Table Exists in Schema
}
