% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_utils.R, R/utils.R
\name{validate_connection}
\alias{validate_connection}
\title{Validate Database Connection}
\usage{
validate_connection(connection)

validate_connection(connection)
}
\arguments{
\item{connection}{Database connection object}
}
\value{
Logical indicating if connection is valid

Logical. TRUE if connection is valid, FALSE otherwise
}
\description{
Validate Database Connection

Validate Database Connection
}
\keyword{internal}
